% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runapp.R
\name{app_results}
\alias{app_results}
\title{Ganaderia Sostenible Web App}
\usage{
app_results(inputs, departamento, municipio)
}
\arguments{
\item{inputs}{List of inputs: a list with year and value for each tipo_cobertura}

\item{departamento}{Department}

\item{municipio}{Municipality}
}
\value{
None
}
\description{
Ganaderia Sostenible Web App
}
\examples{
\dontrun{
runGanaderiaSostenible()
}
}
