% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{availableTipoCobertura}
\alias{availableTipoCobertura}
\title{Available Tipo de Cobertura}
\usage{
availableTipoCobertura()
}
\value{
None
}
\description{
Available Tipo de Cobertura
}
\examples{
availableRegiones()
}
