% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{biodiv_area}
\alias{biodiv_area}
\title{Estimación de carbono capturado en bosques primarios}
\usage{
biodiv_area(area, region, tipo_cobertura, t = 0)
}
\arguments{
\item{area}{área (en hectáreas) implementada por año en cada una de las categorías de uso del suelo}

\item{tipo_cobertura}{Tipo de terreno}

\item{t}{tiempo de estimación}

\item{región}{lugar estimado según Departamento y municipio del predio}
}
\description{
Estimación de carbono capturado en bosques primarios
}
\examples{
biodiv_area(c(1.3, 3, 4), 'Eje Cafetero', 'bosque_primario')

}
