% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_diversity.R
\name{biodiv_area2}
\alias{biodiv_area2}
\title{Bird biodiversity for a given area}
\usage{
biodiv_area2(area, region, tipo_cobertura, t = 0)
}
\arguments{
\item{area}{Area in hectares}

\item{region}{Region}

\item{tipo_cobertura}{Land coverage type}

\item{t}{time}
}
\value{
None
}
\description{
Bird biodiversity for a given area
}
\examples{
biodiv_area2(10, "Eje Cafetero", "bosque_secundario")
}
