% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{cambio_carbono}
\alias{cambio_carbono}
\title{Estimación del cambio en las reservas de carbono según categorías de uso del suelo}
\usage{
cambio_carbono(region, tipo_cobertura, t_f = 0)
}
\arguments{
\item{region}{Nombre de la zona en la cual estan ubicados los terrenos.}

\item{tipo_cobertura}{Categoría de uso del suelo en la cuál tiene ganado.}

\item{t_f}{Tiempo del último año en el que aumento el número de hectareas del terreno}
}
\value{
None
}
\description{
Estimación del cambio en las reservas de carbono según categorías de uso del suelo
}
\examples{
cambio_carbono('Eje Cafetero', 'bosque_secundario', 5)

}
