% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{captura_carbono_bosques}
\alias{captura_carbono_bosques}
\title{Estimación de carbono capturado en bosques primarios}
\usage{
captura_carbono_bosques(departamento, municipio, area_bosque, años, t_e,
  ...)
}
\arguments{
\item{departamento}{Departamento de ubicación del suelo}

\item{municipio}{municipio de ubicación del suelo}

\item{años}{años donde se agregan hectáreas al suelo}

\item{t_e}{tiempo de estimación}

\item{area}{área (en hectáreas) implementada por año en cada una de las categorías de uso del suelo}
}
\description{
Estimación de carbono capturado en bosques primarios
}
\examples{
captura_carbono_bosques('Nariño', 'cordoba', c(12, 324, 211, 89), 2000:2004, 20)

}
