% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{co2_carros}
\alias{co2_carros}
\title{Captured carbon}
\usage{
co2_carros(carbono_capturado)
}
\arguments{
\item{carbono_capturado}{Captured carbon in CO2 tons}
}
\value{
None
}
\description{
Captured carbon
}
\examples{
co2_carros(10000)
}
