% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{factor_emision}
\alias{factor_emision}
\title{Estimación de factores de emisión según categorías de uso del suelo}
\usage{
factor_emision(cb_carbono, region)
}
\arguments{
\item{region}{Nombre de la zona en la cual estan ubicados los terrenos.}

\item{cambio_carbono}{Valor o vector de valores del cambio de carbono según tipo de terreno}
}
\value{
None
}
\description{
Estimación de factores de emisión según categorías de uso del suelo
}
\examples{
factor_emision(c(1.3, 3, 4), 'Eje Cafetero')

}
