% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{regiones_match}
\alias{regiones_match}
\title{Match geographic regions to municipalities}
\usage{
regiones_match(departamento = NULL, municipio)
}
\arguments{
\item{departamento}{Department}

\item{municipio}{Municipality}
}
\value{
None
}
\description{
Match geographic regions to municipalities
}
\examples{
regiones_match(departamento = NULL, municipio = 'usiacuri')
}
