% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{searchInput}
\alias{searchInput}
\title{searchInput}
\usage{
searchInput(inputId, data, placeholder)
}
\arguments{
\item{inputId}{inputId}

\item{data}{data to be searched}

\item{placeholder}{text placeholder for input}
}
\description{
searchInput
}
