% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{viz_bar}
\alias{viz_bar}
\title{Bar visualization of carbon capture}
\usage{
viz_bar(data)
}
\arguments{
\item{data}{Data frame with carbon capture for each land coverage type}
}
\value{
None
}
\description{
Bar visualization of carbon capture
}
\examples{
\dontrun{
viz_bar()
}
}
