% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{viz_lines}
\alias{viz_lines}
\title{Lines visualization of carbon capture}
\usage{
viz_lines(data, type_plot = "spline")
}
\arguments{
\item{data}{Data frame with carbon capture for each land coverage type and year}

\item{type_plot}{Plot type}
}
\value{
None
}
\description{
Lines visualization of carbon capture
}
\examples{
\dontrun{
viz_lines()
}
}
