% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_capture.R
\name{cambio_carbono}
\alias{cambio_carbono}
\title{Carbon capture change}
\usage{
cambio_carbono(captura, region, area = 1)
}
\arguments{
\item{captura}{Vector of carbon emissions changes for each land coverage type}

\item{region}{geographic region name}

\item{area}{area in hectares}
}
\value{
None
}
\description{
Carbon capture change
}
\examples{
cambio_carbono(c(1.3, 3, 4), 'Eje Cafetero')

}
