% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_capture.R
\name{captura_carbono_bosque_primario}
\alias{captura_carbono_bosque_primario}
\title{Captured carbon for forests}
\usage{
captura_carbono_bosque_primario(departamento, municipio, area = 1, t = NULL)
}
\arguments{
\item{departamento}{Department}

\item{municipio}{Municipality}

\item{area}{Hectares of forest}

\item{t}{Time}
}
\value{
None
}
\description{
Captured carbon for forests
}
\examples{
captura_carbono_bosque_primario(departamento = 'Caldas', municipio = 'Manizales',
  area = 500)
}
