% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{captura_carbono_bosques}
\alias{captura_carbono_bosques}
\title{Captured carbon for forests}
\usage{
captura_carbono_bosques(departamento, municipio, area_bosque, anos, t_e)
}
\arguments{
\item{departamento}{Department}

\item{municipio}{Municipality}

\item{area_bosque}{Hectares of forest}

\item{anos}{Years}

\item{t_e}{Estimated times}
}
\value{
None
}
\description{
Captured carbon for forests
}
\examples{
captura_carbono_bosques(departamento = 'Nariño', municipio = 'Córdoba',
  area_bosque = 500, t_e = 1, anos = 2001)
}
