% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_capture_old.R
\name{carbono_capturado_estimacion}
\alias{carbono_capturado_estimacion}
\title{Carbon capture estimation}
\usage{
carbono_capturado_estimacion(area, anos, t_e, region, tipo_cobertura)
}
\arguments{
\item{area}{Area (hectares) of coverage implementation by year for each land coverage type}

\item{anos}{Years}

\item{t_e}{time of estimation}

\item{region}{region}

\item{tipo_cobertura}{Land coverage type}
}
\value{
None
}
\description{
Carbon capture estimation
}
\examples{
pastoriles_area <- c(10.6,	142.5, 0.0,	131.0, 170.0, 222.7, 95.4 )
carbono_capturado_estimacion(pastoriles_area, anos = c(2013:2019), t_e = 7,
  region = 'Bajo Magdalena', tipo_cobertura = 'silvopastoriles')
}
