% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{carbono_capturado_estimacion}
\alias{carbono_capturado_estimacion}
\title{Estimación de factores de emisión según categorías de uso del suelo}
\usage{
carbono_capturado_estimacion(area, t_e, region, tipo_cobertura)
}
\arguments{
\item{area}{área (en hectáreas) implementada por año en cada una de las categorías de uso del suelo}

\item{factor_emision}{Cambio (t CO2e ha-1) por tipo de suelo}
}
\value{
None
}
\description{
Estimación de factores de emisión según categorías de uso del suelo
}
\examples{
carbono_capturado(area = c(6418.8, 43, 0, 69.5), factor_emision = c(-0.4, 2.3, 3.5, 4.4))

}
