% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{carbono_capturado_estimacion}
\alias{carbono_capturado_estimacion}
\title{Estimación de carbono capturado}
\usage{
carbono_capturado_estimacion(area, años, t_e, region, tipo_cobertura)
}
\arguments{
\item{area}{área (en hectáreas) implementada por año en cada una de las categorías de uso del suelo}

\item{años}{años donde se agregan hectáreas al suelo}

\item{t_e}{tiempo de estimación}

\item{region}{Lugar de ubicación del suelo}

\item{tipo_cobertura}{Categoría de uso del suelo en la cuál tiene ganado.}
}
\value{
None
}
\description{
Estimación de carbono capturado
}
\examples{
carbono_capturado_estimacion(area = c(6418.8, 43, 0, 69.5), t_e = 20, region = 'Eje Cafetero', tipo_cobertura = 'cercas_vivas')

}
