% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{co2_carros}
\alias{co2_carros}
\title{Estimación de carros por carbono capturado}
\usage{
co2_carros(carbono_capturado)
}
\arguments{
\item{cambio_carbono}{Valor del total de carbono según tipo de terreno}
}
\value{
None
}
\description{
Estimación de carros por carbono capturado
}
\examples{
co2_carros(3425)

}
