% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_capture.R
\name{estimacion_co2}
\alias{estimacion_co2}
\title{Carbon capture estimation in tidy format}
\usage{
estimacion_co2(inputs, departamento, municipio, this_year = NULL,
  t_max = 20)
}
\arguments{
\item{inputs}{List with implementations by year and value of hectares}

\item{departamento}{Departamento}

\item{municipio}{Municipio}
}
\value{
None
}
\description{
Carbon capture estimation in tidy format
}
\examples{
inputs <- list(
  bosque_primario = list(year = current_year - 8, value = 5),
  bosque_secundario = list(year = current_year - 8, value = 10),
  arboles_dispersos = list(year = current_year - 8, value = 2),
  cercas_vivas = list(year = current_year - 8, value = 2),
  silvopastoriles = list(year = current_year - 8, value = 5)
)
estimacion_carbono(inputs, departamento = "Quindio", municipio = "Montenegro")
}
