% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_capture.R
\name{estimacion_co2_tidy}
\alias{estimacion_co2_tidy}
\title{Carbon capture estimation}
\usage{
estimacion_co2_tidy(inputs, departamento, municipio, t_max = 20)
}
\arguments{
\item{inputs}{List with implementations by year and value of hectares}

\item{departamento}{Departamento}

\item{municipio}{Municipio}

\item{region}{region}

\item{tipo_cobertura}{Land coverage type}
}
\value{
None
}
\description{
Carbon capture estimation
}
