% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_capture_old.R
\name{factor_emision}
\alias{factor_emision}
\title{Carbon Emision Factor}
\usage{
factor_emision(cb_carbono, region)
}
\arguments{
\item{cb_carbono}{Vector of carbon emissions changes for each land coverage type}

\item{region}{geographic region name}
}
\value{
None
}
\description{
Carbon Emision Factor
}
\examples{
factor_emision(c(1.3, 3, 4), 'Eje Cafetero')

}
