% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{regiones_match}
\alias{regiones_match}
\title{Región}
\usage{
regiones_match(departamento = NULL, municipio, ...)
}
\arguments{
\item{departamento}{Departamento}

\item{municipio}{Municipio}
}
\value{
None
}
\description{
Región
}
\examples{
regiones_match('Boyaca', 'Caldas')

}
