% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runapp.R
\name{runGanaderiaSostenible}
\alias{runGanaderiaSostenible}
\title{Ganaderia Sostenible Web App}
\usage{
runGanaderiaSostenible()
}
\value{
None
}
\description{
Ganaderia Sostenible Web App
}
\examples{
\dontrun{
runGanaderiaSostenible()
}
}
